test(onboard): cover cleanup before fatal exit
